+++ generator_tests/test/doc/login_nullable_test.dart
@@ -1,5 +1,4 @@
 @Timeout(Duration(seconds: 145))
-
 import 'package:test/test.dart';
 
 import '../helpers.dart';
@@ -39,13 +38,13 @@ void main() {
               final double? height;
             
               LoginNullable({
-                @FormControlAnnotation() this.email,
-                @FormControlAnnotation() this.password,
-                @FormControlAnnotation() this.rememberMe,
-                @FormControlAnnotation() this.theme,
-                @FormControlAnnotation() this.mode,
-                @FormControlAnnotation() this.timeout,
-                @FormControlAnnotation() this.height,
+                @FormControlAnnotation<String>() this.email,
+                @FormControlAnnotation<String>() this.password,
+                @FormControlAnnotation<bool>() this.rememberMe,
+                @FormControlAnnotation<String>() this.theme,
+                @FormControlAnnotation<UserMode>() this.mode,
+                @FormControlAnnotation<int>() this.timeout,
+                @FormControlAnnotation<double>() this.height,
               });
             }
           ''',