+++ generator_tests/test/doc/user_profile_test.dart
@@ -1,5 +1,4 @@
 @Timeout(Duration(seconds: 145))
-
 import 'package:test/test.dart';
 
 import '../helpers.dart';
@@ -33,17 +32,17 @@ void main() {
               final String lastName;
             
               final Address home;
-            
+                        
               final Address? office;
             
               UserProfile({
                 required this.id,
-                @FormControlAnnotation(
-                  validators: const [requiredValidator],
+                @FormControlAnnotation<String>(
+                  validators: [requiredValidator],
                 )
                     this.firstName = '',
-                @FormControlAnnotation(
-                  validators: const [requiredValidator],
+                @FormControlAnnotation<String>(
+                  validators: [requiredValidator],
                 )
                     this.lastName = '',
                 required this.home,
@@ -60,13 +59,13 @@ void main() {
               final String? zip;
             
               Address({
-                @FormControlAnnotation()
+                @FormControlAnnotation<String>()
                     this.street,
-                @FormControlAnnotation(
-                  validators: const [requiredValidator],
+                @FormControlAnnotation<String>(
+                  validators: [requiredValidator],
                 )
                     this.city,
-                @FormControlAnnotation()
+                @FormControlAnnotation<String>()
                     this.zip,
               });
             }
@@ -436,14 +435,18 @@ class UserProfileForm implements FormModel<UserProfile> {
   FormGroup formElements() => FormGroup({
         firstNameControlName: FormControl<String>(
             value: userProfile.firstName,
-            validators: [requiredValidator],
+            validators: [
+              (control) => requiredValidator(control as FormControl<String>)
+            ],
             asyncValidators: [],
             asyncValidatorsDebounceTime: 250,
             disabled: false,
             touched: false),
         lastNameControlName: FormControl<String>(
             value: userProfile.lastName,
-            validators: [requiredValidator],
+            validators: [
+              (control) => requiredValidator(control as FormControl<String>)
+            ],
             asyncValidators: [],
             asyncValidatorsDebounceTime: 250,
             disabled: false,
@@ -677,7 +680,9 @@ class AddressForm implements FormModel<Address> {
             touched: false),
         cityControlName: FormControl<String>(
             value: address?.city,
-            validators: [requiredValidator],
+            validators: [
+              (control) => requiredValidator(control as FormControl<String>)
+            ],
             asyncValidators: [],
             asyncValidatorsDebounceTime: 250,
             disabled: false,