+++ packages/reactive_forms_generator/example/lib/docs/mailing_list/mailing_list.dart
@@ -10,7 +10,7 @@ class MailingList {
   final List<String?> emailList;
 
   MailingList({
-    @FormArrayAnnotation(
+    @FormArrayAnnotation<String>(
       validators: [
         mailingListValidator,
       ],