+++ packages/reactive_forms_generator/lib/src/form_elements/form_control_generator.dart
@@ -11,15 +11,6 @@ class FormControlGenerator extends FormElementGenerator {
 
   @override
   String element() {
-    final props = [
-      'value: $value',
-      'validators: [${syncValidatorList(formControlChecker).join(',')}]',
-      'asyncValidators: [${asyncValidatorList(formControlChecker).join(',')}]',
-      'asyncValidatorsDebounceTime: ${asyncValidatorsDebounceTime(formControlChecker)}',
-      'disabled: ${disabled(formControlChecker)}',
-      'touched: ${touched(formControlChecker)}',
-    ].join(',');
-
     String displayType = field.type.getDisplayString(withNullability: true);
 
     // we need to trim last NullabilitySuffix.question cause FormControl modifies
@@ -27,6 +18,43 @@ class FormControlGenerator extends FormElementGenerator {
     if (field.type.nullabilitySuffix == NullabilitySuffix.question) {
       displayType = displayType.substring(0, displayType.length - 1);
     }
+
+    final annotationType =
+        (annotation(formControlChecker)?.type as ParameterizedType)
+            .typeArguments
+            .first
+            .toString();
+
+    if (annotationType != 'dynamic' && annotationType != displayType) {
+      throw Exception(
+        'Annotation and field type mismatch. Annotation is typed as `$annotationType` and field(`${field.name}`) as `$displayType`.',
+      );
+    }
+
+    List<String> validators = syncValidatorList(formControlChecker);
+    List<String> asyncValidators = asyncValidatorList(formControlChecker);
+
+    if (annotationType != 'dynamic') {
+      validators = validators
+          .map(
+            (e) => '(control) => $e(control as FormControl<$displayType>)',
+          )
+          .toList();
+      asyncValidators = asyncValidators
+          .map(
+            (e) => '(control) => $e(control as FormControl<$displayType>)',
+          )
+          .toList();
+    }
+
+    final props = [
+      'value: $value',
+      'validators: [${validators.join(',')}]',
+      'asyncValidators: [${asyncValidators.join(',')}]',
+      'asyncValidatorsDebounceTime: ${asyncValidatorsDebounceTime(formControlChecker)}',
+      'disabled: ${disabled(formControlChecker)}',
+      'touched: ${touched(formControlChecker)}',
+    ].join(',');
 
     return 'FormControl<$displayType>($props)';
   }