+++ packages/reactive_forms_annotations/CHANGELOG.md
@@ -1,38 +1,59 @@
+## [0.7.0-beta]
+
+* FormArrayAnnotation with generic typing
+
+## [0.6.0-beta]
+
+* FormControlAnnotation with generic typing
+
 ## [0.5.0-beta]
+
 * ExtendedControl model
 
 ## [0.4.0-beta]
+
 * reactive_forms bump
 
 ## [0.3.2-beta]
+
 * name parameter added
 
 ## [0.3.1-beta]
+
 * FormModel interface
 
 ## [0.3.0-beta]
+
 * rf11
 
 ## [0.2.0-beta]
+
 * package update
 
 ## [0.1.1-beta]
+
 * package update
 
 ## [0.1.0-beta]
+
 * package update
 
 ## [0.0.5-beta]
+
 * extended FormArray annotation
 
 ## [0.0.4-beta]
+
 * asyncDebounceTime and disable support
 
 ## [0.0.3-beta]
+
 * FormGroup support
 
 ## [0.0.2-beta]
+
 * small fix
 
 ## [0.0.1-beta]
+
 * initial release.